creates sphinx documentation
with this, we set up sphinx for generating documentation. currently,
there's basically nothing here (just the docstrings), but now we have
it and can add as we go.

we won't be able to have anything on readthedocs.org until we make the
repo public, but once we do, we should be able to set it up pretty
quickly (I've added a .readthedocs.yml that we should be able to use
as is)