Corrects pooling calculation
with this commit, we correct the pooling calculation. we were
previously averaging the values of the model statistics in pooling
windows by summing across space and dividing by the number of non-zero
pixels (defined using some zero_thresh). what we should have been
doing is to divide by the sum of window values. our binarized
approximation of this was over-weighting large windows. this also
makes the values in our representation larger, which may help things?

 - removes zero_thresh arg and window_num_pixels attribute, since we
   no longer need them

 - PoolingWindows.forward now does both the windowing and the pooling,
   which are split into two functions: window() and pool(). these can
   be called separately if the user wishes. all of forward(),
   window(), and pool() except either appropriately sized tensors or
   dicts containing those appropriately sized tensors

 - this cleans up the VentralStream models forward() code
   significantly (though a bit of work is still necessary to make sure
   they're the right shape)