PoolingWindows can now deal with images of arbitrary size
with this, PoolingWindows are able to handle images of any size. we
still have to create the windows as a square, so we create them too
big and then slice them down to the proper size (appropriately
centered). For images that aren't a power of 2, it should work the
same way as the steerable pyramid and other pyramid methods