overhaul to how we handle dimensionality
we return to the VentralStream models' windows always being
3d (instead of 4d). we do this for a couple reasons: when the image is
not square, some of those windows will be empty and it's a waste to
carry them around, but if we want to keep the 4d structure, we need
to. this way we can always assume that PoolingWindows.windows is a
list of 3d tensors, which makes our lives easier

another big change: we no longer insert nans between eccentricity
bands when doing VentralStream._representation_for_plotting. this
turned out to be too much work when we don't have that structure and,
in practice, isn't actually all that helpful. the clean_stem_plot
function can still make those plots, but it's not being used
anymore. this means _representation_for_plotting no longer returns
xvals

other changes:

 - to_numpy() call no longer happens in
   Metamer.representation_ratio(). it does happen in
   _representation_for_plotting. basically to_numpy() should only
   happens when we know we're about to plot. therefore, also happens
   in rescale_ylim function

 - removes RetinalGanglionCells._representation_for_plotting, the
   super version is fine for now (PrimaryVisualCortex still keeps it
   though, because it goes from 3d tensor to dict of 3d tensors)