adds normalization for PrimaryVisualCortex
we want to add the ability to normalize the PrimaryVisualCortex
model's representation. there's now a boolean arg during
initialization: if False, we do as we do now, no normalization; if
True, we make sure that each of the values in the representation dict
have an L2-norm of 1

adds two new functions to accompany this: normalize and
normalize_dict (first acts on a single tensor, the second on a
dictionary of tensors).

instead of just checking if a key is tuple in
rectangular_to_polar_dict and locaL_gain_control, we don't add it to
the returned dictionaries if its not a tuple and doesn't starts with
'residual'. this way we can operate on more general dictionaries

also cleans up the documentation of the other functions in
non_linearities.py a bit