adds normalization
adds two new helper functions to non_linearities: generate_norm_stats
and zscore_stats. generate_norm_stats takes a model and a directory,
runs the model on every image in that directory, and saves the
resulting statistics. if the model has a to_normalize attribute, we
save those attributes, otherwise we save the value returned by
model(images).

this is to be used with the zscore_stats function, which takes a model
and the dictionary generated above, and zscores them (using the mean
and std across images). caveat here is there's no way I can abstract
away when it should be called, so that's up to the user