Adds angle/eccen_index to pooling windows
PoolingWindows creation is fairly memory-intensive, especially for
small scaling values / large images, and it uses more memory than
either the finished PoolingWindows object or the use of that
object. in order to try and get around that, we provide the use of
angle_index and eccen_index, which allow the user to create the
windows in pieces (each run of which will take relatively little time
and memory). not present here is a way to figure out which index
values to use or how to concatenate them correctly; I have that
somewhere externally