MAJOR BUGFIX: V1 model properly adjusts mean luminance
since the V1 model's mean luminance was coming from `self.image`,
instead of `image`, which is what was getting adjusted (and what the
complex pyramid was built on, as well as the RGC representation), the
optimizer could not adjust the image to change the mean luminance
representation -- it wasn't included in the graph. this fixes that.