Adds coarse-to-fine optimization
adds ability to do optimization in coarse to fine order, as they do in
the Portilla Simoncelli paper. There's a bit of an abstraction of this
in the metamer class, which requires two additional features from the
model: a scales attribute, which lists the scales in *reverse* order,
and the ability for the forward method to accept a scales keyword
argument. currently only VentralStream.PrimaryVisualCortex has this
ability

should work with the previously-added fraction_removed and loss_change
options

also adds tests