improves speed of pooling windows
on GPUs, the indexing operation I was doing with the window_size_mask
took a lot of time (about twice as much as the einsum), so this gets
rid of that. now, instead, we just clamp window_sizes to have a
minimum of 1 (the value I was using as my cut-off), and we do that
at their creation during initialization

this speeds things up quite a bit (and more so for smaller scaling
values / larger images). also removed the calls to empty_cache becuase
it's not clear to me that was actually helping with my memory issues
and that was also costing serious time

this also has the positive side-effect of cleaning up a fair amount of
the code in the forward, pool, and project functions