Adds initial MAD competition implementatin
with this commit, we add an initial implementation of MAD
competition. it currently only minimizes model 1 while holding model 2
constant, and only works on models (not metrics), but the basic
framework is there. a fair amount of cleanup is required as well.

MAD competition (like Metamer) inherits the Synthesis class and so
makes use of much there. more functionality consolidation is needed
here, as there was a lot of copy-and-paste

 - Synthesis now uses the self.names dictionary to determine what to
   target / match etc. this doesn't change how Metamer works (or
   honestly, probably any of the other synthesis methods) but is
   necessary for MAD

 - adds a base gradient descent optimizer choice, GD

 - scheduler is now optional

 - optimizer_step's pbar arg is now optional. if None, will not update
   information. can also add additional information to the pbar as
   kwargs