mad competition now supports metrics as well as models
the MAD competition paper, as originally written, compared SSIM and
MSE, which are metrics, not models (they take two images and return a
scalar distance between them). as our code was written, it only worked
for models (which take an image and return a representation, and we
use the L2-norm of the difference between representations for the
distance between images).

with this commit, can now use metrics for MAD competition. to do that,
we construct a 'dummy model', that just returns a copy of the image
and modify the function called for our objective function

other changes

 - adds Identity() class in simulate/models/naive.py

 - adds MSE class to mse() function, since metrics should be functions
   now

 - we no longer do the whole randomizing of pixels before computing
   the loss in Synthesis._closure(). this really messes up those
   metrics that extract something meaningful from the image (i.e., all
   the ones we're interested in). so this is turned off by default,
   and only turned on when needed. right now, this is just when
   metamer has fraction_removed > 1 or loss_change_fraction <
   1. should try and generalize that tools

 - updates docstrings

 - adds loss_function attr, which is how we switch things around for
   the two models. objective_function calls this on x and y. (still
   use objective_function because it handles the normalizing)

 - adds _get_model_name, which first tries to grab the model's name
   attribute and, if it doesn't have one, then returns the class
   name. useful for the Identity class

 - raise a warning whenever the representation error is computed with
   a metric, since it will be meaningless