adds piecewise log for new way of handling fovea
this commit starts a new way of handling the fovea for the pooling
windows: use a piecewise log transform, which is linear in some region
and log everywhere else. this allows us to build windows all the way
down to the fovea and have therefore have them tile the entire image
uniformly (we'll drop a mask on it outside of this to exclude the
fovea and make it a pixel-match area). this currently hasn't been
updated for PoolingWindows class yet, but will be used for the DoG
windows