removes msssim, SSIM, MSSSIM, restructures ssim
with this commit, we remove msssim (because there are some tricky things
that I don't have the time to figure out -- issue opened for it if
people want to check it out) and SSIM (caching window no longer makes
sense), and thus MSSSIM

also restructures ssim so that we have a helper function which computes
the ssim_map, contrast map, and weight. ssim just returns the mean over
the image. adds new function, ssim_map, which just returns the map.
contrast map will be used by msssim if we re-implement it