removes PoolingWindows parallel support
We were supporting the ability to parallelize PoolingWindows (and thus
PooledRGC and PooledV1) across GPUs, but this removes it. I was no
longer using it and we weren't testing it (despite having tests for it,
we never had access to a single GPU, let alone multiple GPUs, for
testing).

The reason I stopped using this is that, during metamer synthesis, there
were obvious discontinuities between the sections of angular windows
that were on different GPUs, which I think is due to the fact that it's
hard to set the state of the RNG identically across GPUs (though support
for this in pytorch is improving). Therefore, my suggestions is to just
make your model small enough to fit on a single GPU instead of messing
around with multiple ones.