fix for animate tests
apparently, exceptions raised in the movie_plot function get printed
when animate() is run but don't get raised until the animation is saved
or converted in some way, which meant that we had failing tests that
looked like they were passing. with this, we make sure to call
anim.to_html5_video(), which will cause exceptions to be raised. this
also requires adding ffmpeg as an additional requirement for tests