synthesis plotting code works with RGB/A images
allows user to set channel_idx=None everywhere we had previously
required it to be an int, which enables them to work. for
plot_synthesis_status, if user does this we set as_rgb=True (because we
assume we're only plotting a single image).

this swaps out pt.imshow for po.imshow