removes fraction_removed and loss_change_fraction
removes support for two synthesis options: fraction_removed and
loss_change_fraction. together, they supported the ability to only
compute gradient with respect to some subset of representation if it
looked like loss had stopped changing.

this was not being used by anyone and was getting more complicated to
support (my initial implementation assumed only one batch and channel
in the representation and failed if that wasn't true)

loss_change_thresh and loss_change_iter remain, because they're used for
coarse-to-fine