overhauls save/load
save and load no longer handle models. instead, load is now an instance
method (not a class method) and so the object must be initialized first.
we then call load, modifying it in place. we check whether some
attributes (those set on initialization) are equal and raise an
exception if not

also:

 - loss_1/2_norm should be set in MADCompetition's initialization

 - removed some unnecessary calls to plot and animate in test_mad