bugfix and future-proof in load_images
image #84: starting in scikit-image 0.19, rgb2gray will treat 2d signals
as 1d images with 3 color channels. this makes sure we only call
rgb2gray on signals with more than 2 dimensions, to avoid that

we also only rearrange the axes when as_gray=False if the signal has
more than 2 dimensions, to avoid swapping the height and width of a
grayscale image. this was a problem in the Demo_Eigendistortion
notebook (not sure why it just became noticeable in this branch)