properly init fail model
this test shouldn't have passed before -- the model was not properly
initialized and thus it was failing for the wrong reason. now, it was
failing during the initialization of Metamer because I compute the first
loss, and thus it was apparent that model wasn't initialized. With this,
it fails correctly