Adds necessary stats mask
This creates a dictionary and mask which mark the redundancies in the
auto- and cross-correlation statistics. Follows the same logic as the
scales_shape_dict: it should be a dictionary whose keys are the
statistic names and whose values are tensors with the same shape as
the computed stat (except for batch and channel).

this then requires changing convert_to_dict to make use of this mask,
placing NaNs where the redudant stats are.