change shape of auto_corr_mag and mags_std
in previous commit, both of these tensors had orientations along the
final dimension and scales along the second to last. this commit swaps
those two

this makes mags_std the same as before these changes, but swaps the
final two of auto_corr_mag. worth doing because it makes it consistent
with the other statistics, which always have scales along the final dim