PS now accepts weights argumnet
with this commit, we now have a single PS class which accepts a
weights argument. If None (default), then we compute stats across
whole image. else, we use them as weighting regions. in both cases,
the output is now 4d, with the third dimension holding the number of
weighting regions, which is a singleton if weights=None

to accomplish this, we make use of the WeightedAverage class, as well
as new helper modules: StatsComputer and WeightedComputer, which
handle the computations for pixel_stats, autocorr, cross-corr, and
skew/kurt_recon.

docstrings need to be updated / written, everything needs to be tested
a bit more (and new tests need to be written), and it's possible
things can be made simpler / more efficient, but I believe the basics
are now working