moves DemoEigendsitortion synth to tests
with this, we move the synthesis in Demo_Eigendistortion notebook into a
regression test instead.

the call for this synthesize still displayed in the notebook (in an
admonition), but not ran. we also add a new linter which ensures the
synthesize calls are the same in the notebook and the test