optimization: moving power calculation out of loop
Based on a cursory look, it seems that the self.order property is not
mutated after initialization: if this is the case, then we do not need
to recompute the complex_const within each loop, but can instead compute
it once on top of the loop, and thus reduce calculations performed.

According to profiling, the use of `{method 'pow' of
'torch._C.TensorBase' objects}` may account for as much as a fifth of
the runtime of `forward` and `_recon_levels`. If these uses of
`np.power()` in the code do in fact correspond to that profile line,
this could lead to a nontrivial speedup of both functions.